clock_font_img/pyproject.toml
[project]
name = "clock-font-img"
version = "0.1.0"
description = "Generate clock font images using MiSans font"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=10.0.0",
    "PyInquirer>=1.0.3",
]
